public/solana_bots/measure-token-performance/requirements.txt
asyncio==3.4.3
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
requests==2.32.3
rsa==4.9
soupsieve==2.5
telegram_text==0.2.0
Telethon==1.36.0
urllib3==2.2.2
schedule==1.2.2
pandas==2.2.2